requirements.txt
certifi==2024.12.14
elastic-transport==8.17.0
elasticsearch==8.17.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
six==1.17.0
urllib3==2.3.0
